Validation on all inputs

+++ modifyTask.php
@@ -115,7 +115,7 @@ if(mysqli_num_rows($result1)>0){
     <input type="submit" name="Change_Leader" value="Change Leaders">
     <input type="submit" name="Change_Members" value ="Change Members">
 <input type="submit" name="load" value="Change"><br>
-
+<hr>
 
 </form>
 
@@ -165,17 +165,25 @@ if(isset($_POST["load"])){ //loading info into the change textboxes and changing
 
 if(isset($_POST["edit"])){ //update info of tasks in database
 
-    
-
     $task_id = $_SESSION["task_id"];
+
+    $sql="SELECT Task_title,priority,Content FROM tasks WHERE task_id = ".$task_id;
+    if($result= mysqli_query($conn,$sql)){
+        if(mysqli_num_rows($result)>0){
+            while($row = mysqli_fetch_array($result)){
+                $addToHistory="INSERT INTO tasks_history(task_id, task_title, Content, priority) 
+                VALUES(".$task_id.",".$row["Task_title"].", ".$row["Content"].",".$row["priority"].")";
+                $query2= mysqli_query($conn,$addToHistory);
+                echo mysqli_error($conn); 
+            }
+        }
+    }
+
     $title= mysqli_real_escape_string($conn,$_POST['TaskTitle']);
     $info = mysqli_real_escape_string($conn, $_POST['TaskInfo']);
     $priority = mysqli_real_escape_string($conn,$_POST["priority"]);
     
-    $addToHistory="INSERT INTO tasks_history(task_id, task_title, Content, priority) 
-    VALUES('$task_id','$title','$info','$priority')";
-    $query2= mysqli_query($conn,$addToHistory);
-    echo mysqli_error($conn); 
+    
     $addTask = "UPDATE tasks SET Task_title ='$title' ,Content = '$info',priority = '$priority' 
     WHERE task_id = '$task_id'";
     $query = mysqli_query($conn, $addTask);
@@ -194,7 +202,10 @@ if(isset($_POST["Change_Leader"])){
         alert("Please choose task to edit")
         </script>';
     }elseif(empty($_POST['check_list_leaders'])){
-
+        echo '<script>
+        alert("Please choose new leaders")
+        </script>';
+    }else{
     
         $task_id = mysqli_real_escape_string($conn,$_POST['chosen_task']);
         $leaderslist= $_POST["check_list_leaders"];
@@ -214,16 +225,25 @@ if(isset($_POST["Change_Leader"])){
         echo "<meta http-equiv='refresh' content='0'>";
 
 
-
     }
+    
 }
 
 if(isset($_POST["Change_Members"])){ //adding button to change members
 
-    $task_id = mysqli_real_escape_string($conn,$_POST['chosen_task']);
-    $leaderslist= $_POST["check_list_leaders"];
+    if(empty($_POST['chosen_task'])){
+        echo '<script>
+        alert("Please choose task to edit")
+        </script>';
+    }elseif(empty($_POST['check_list_leaders'])){
+        echo '<script>
+        alert("Please choose new leaders")
+        </script>';
+    }else{
 
-    // if($task_id && $leaderslist){        condition to handle if no inputs inserted
+        $task_id = mysqli_real_escape_string($conn,$_POST['chosen_task']);
+        $leaderslist= $_POST["check_list_leaders"];
+
         $removeTask = "DELETE FROM task_members WHERE task_id = ".$task_id;
         $query = mysqli_query($conn, $removeTask);
 
@@ -240,13 +260,19 @@ if(isset($_POST["Change_Members"])){ //adding button to change members
 
 
 
-    // }
+    }
 }
 
 
 
 //change due date form
 if(isset($_POST["ChangeDate"])){
+
+    if(empty($_POST['tasks_list'])){
+        echo '<script>
+        alert("Please choose task to edit")
+        </script>';
+    }else{
 
     //task id from table
         $task_id= mysqli_real_escape_string($conn,$_POST['tasks_list']);
@@ -276,6 +302,8 @@ if(isset($_POST["ChangeDate"])){
         mysqli_query($conn,$updateDue);
         
     }
+}
+
 ?>
 <form action="<?php echo $_SERVER['PHP_SELF']?>" method="post">
 
@@ -313,7 +341,7 @@ if($result= mysqli_query($conn,$sql)){
 <h1>Change Task Due Date</h1>
 
 <label for="new_due">New Due Date</label>
-<input type="date" name="new_due">
+<input type="date" name="new_due" required>
 <input type="submit" name="ChangeDate">
 </form>
